New contract between core and plugins
A plugin configuration consists of the plugin (the name of a package or
the initialization function itself) and a bucket name. The bucket name
is the kind of output that is created by the pipeline.

The initialization function that receives parameters to initialize it.
It returns a function that will run this specific pipeline (either for
all or for a list of files).

Plugins run in groups (the above mentioned buckets).