Use minimist features
In minimist, boolean flags should be marked as such. An example:

foo --bar abc

will result in { bar: "abc" } if `bar` is not marked as being a boolean.
If it is marked as a boolean, `bar` will be set to true and `abc` will
be provided as an additional argument in an array.

minimist also supports defaults. Why not use them?